Fix warnings (#291)
* Fix warnings on windows

* Update event.cpp

* Update input.cpp

* Fix utf8 warnings

* Fix utf8 warnings

* Suppress warning

* Try to show warnings in PR

+++ cpp-terminal/platforms/cursor.cpp
@@ -12,7 +12,7 @@ Term::Cursor Term::cursor_position()
 {
 #if defined(_WIN32)
   CONSOLE_SCREEN_BUFFER_INFO inf;
-  if(GetConsoleScreenBufferInfo(Private::out.handle(), &inf)) return Term::Cursor(inf.dwCursorPosition.Y + 1, inf.dwCursorPosition.X + 1);
+  if(GetConsoleScreenBufferInfo(Private::out.handle(), &inf)) return Term::Cursor(static_cast<std::size_t>(inf.dwCursorPosition.Y + 1), static_cast<std::size_t>(inf.dwCursorPosition.X + 1));
   else
     return Term::Cursor(0, 0);
 #else
